Fix perm error for serve_File from a multi-user directory
When serve_file used on a file that is located in a multi-user file directory,
the current implementation failed because it tried to glob all the content,
and then serve the singe file with a new file server instance.

With the new change, the single file serving no longer use Dir.glob,
and only focuses on serving the single file.